+++ src/.internal/core/util/Animation.ts
@@ -113,9 +113,14 @@ export function getInterpolate(from: any, to: any): (diff: Time, from: any, to: 
 }
 
 
+export enum AnimationState {
+	Stopped,
+	Playing,
+	Paused,
+}
+
 export interface IAnimation {
-	// Returns true if the animation is still playing, or false if the animation isn't playing
-	_runAnimation(_currentTime: number): boolean;
+	_runAnimation(_currentTime: number): AnimationState;
 }
 
 
@@ -160,11 +165,11 @@ export class AnimationTime implements IAnimation {
 		}
 	}
 
-	public _runAnimation(currentTime: number): boolean {
+	public _runAnimation(currentTime: number): AnimationState {
 		if (this._playingDuration !== null) {
 			if (this._startingTime === null) {
 				this._startingTime = currentTime;
-				return true;
+				return AnimationState.Playing;
 			}
 			else {
 				const diff = (currentTime - this._startingTime) / this._playingDuration;
@@ -180,7 +185,7 @@ export class AnimationTime implements IAnimation {
 					if (this.events.isEnabled(type)) {
 						this.events.dispatch(type, { type: type, target: this });
 					}
-					return false;
+					return AnimationState.Stopped;
 
 				} else {
 					this._current = range(diff, this._from as number, this._to as number);
@@ -191,12 +196,12 @@ export class AnimationTime implements IAnimation {
 						this.events.dispatch(type, { type: type, target: this, progress: diff });
 					}
 
-					return true;
+					return AnimationState.Playing;
 				}
 			}
 
 		} else {
-			return false;
+			return AnimationState.Stopped;
 		}
 	}