+++ src/.internal/core/util/Entity.ts
@@ -5,7 +5,7 @@ import type { ILocale } from "./Language";
 
 import { IDisposer, Disposer } from "./Disposer";
 import { EventDispatcher, Events } from "./EventDispatcher";
-import { Time, IAnimation, getInterpolate } from "./Animation";
+import { Time, IAnimation, AnimationState, getInterpolate } from "./Animation";
 import { States } from "./States";
 import { registry } from "../Registry";
 
@@ -388,7 +388,6 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 	protected _animatingSettings: Animated<this["_settings"]> = {};
 	protected _animatingPrivateSettings: Animated<this["_privateSettings"]> = {};
 
-	private _playingAnimations: number = 0;
 	private _disposed: boolean = false;
 
 	// TODO move this into Entity
@@ -417,13 +416,18 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 	 */
 	public abstract markDirty(): void;
 
-	public _runAnimation(currentTime: number): boolean {
+	public _runAnimation(currentTime: number): AnimationState {
+		let state = AnimationState.Stopped;
+
 		if (!this.isDisposed()) {
+			let playing = false;
+			let paused = false;
+
 			$object.each(this._animatingSettings, (key, animation) => {
-				if (animation._stopped) {
+				if (animation.stopped) {
 					this._stopAnimation(key);
 
-				} else if (animation._playing) {
+				} else if (animation.playing) {
 					animation._run(currentTime);
 
 					const diff = animation.percentage;
@@ -431,22 +435,28 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 					if (diff >= 1) {
 						if (animation._checkEnded()) {
 							this.set(key, animation._value(1));
+
 						} else {
+							playing = true;
 							animation._reset(currentTime);
 							this._set(key, animation._value(1));
 						}
 
 					} else {
+						playing = true;
 						this._set(key, animation._value(diff));
 					}
+
+				} else {
+					paused = true;
 				}
 			});
 
 			$object.each(this._animatingPrivateSettings, (key, animation) => {
-				if (animation._stopped) {
+				if (animation.stopped) {
 					this._stopAnimationPrivate(key);
 
-				} else if (animation._playing) {
+				} else if (animation.playing) {
 					animation._run(currentTime);
 
 					const diff = animation.percentage;
@@ -456,21 +466,30 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 							this.setPrivate(key, animation._value(1));
 
 						} else {
+							playing = true;
 							animation._reset(currentTime);
 							this._setPrivate(key, animation._value(1));
 						}
 
 					} else {
+						playing = true;
 						this._setPrivate(key, animation._value(diff));
 					}
+
+				} else {
+					paused = true;
 				}
 			});
 
-			return this._playingAnimations !== 0;
+			if (playing) {
+				state = AnimationState.Playing;
 
-		} else {
-			return false;
+			} else if (paused) {
+				state = AnimationState.Paused;
+			}
 		}
+
+		return state;
 	}
 
 	public abstract _startAnimation(): void;
@@ -639,7 +658,6 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 		const animation = this._animatingSettings[key];
 
 		if (animation) {
-			--this._playingAnimations;
 			delete this._animatingSettings[key];
 			animation.stop();
 		}
@@ -754,7 +772,6 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 		const animation = this._animatingPrivateSettings[key];
 
 		if (animation) {
-			--this._playingAnimations;
 			animation.stop();
 			delete this._animatingPrivateSettings[key];
 		}
@@ -826,8 +843,6 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 
 				const animation = this._animatingSettings[key] = new Animation(this, from, to, duration, easing, loops, this._animationTime());
 
-				++this._playingAnimations;
-
 				this._startAnimation();
 
 				return animation;
@@ -862,8 +877,6 @@ export abstract class Settings implements IDisposer, IAnimation, IStartAnimation
 				this.setPrivate(key, from);
 
 				const animation = this._animatingPrivateSettings[key] = new Animation(this, from, to, duration, easing, loops, this._animationTime());
-
-				++this._playingAnimations;
 
 				this._startAnimation();